CommandAction index to 32 bit

+++ terminalCommands/terminalCommands.go
@@ -31,8 +31,8 @@ type TerminalResponse struct {
 // CommandAction ---
 type CommandAction struct {
 	Id    string `json:"id"`
-	Index uint16 `json:"index"`
-	Act   uint8  `json:"act"`
+	Index uint32 `json:"index"`
+	Act   uint32 `json:"act"`
 	Ton   uint32 `json:"ton"`
 	Toff  uint32 `json:"toff"`
 }
